Field class extracted from App.js
